pyproject.toml
[tool.poetry]
name = "discord-bot-heroku"
version = "0.7.5"
description = ""
authors = ["tetsuya-ki <64536338+tetsuya-ki@users.noreply.github.com>"]
license = "MIT License"

[tool.poetry.dependencies]
python = "^3.10"
"discord.py" = "^2.1.0"
python-dotenv = "^0.20.0"
mojimoji = "^0.0.12"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
